fix(auth): allow SignUpForm to render children
